[Modify] storage 클라이언트에서만 사용되도록 처리

+++ src/utils/storage.ts
@@ -1,46 +1,50 @@
+// utils/storage.ts
+
+const isClient = typeof window !== 'undefined';
+
 const storage = {
   session: {
     get(key: string) {
+      if (!isClient) return null;
       return key ? window.sessionStorage.getItem(key) : window.sessionStorage;
     },
     set(data: Record<string, any> = {}) {
+      if (!isClient) return;
       const keys = Object.keys(data);
-      if (keys.length > 0) {
-        keys.forEach(key => {
-          const value = data[key];
-          window.sessionStorage.setItem(key, value !== undefined ? value : "");
-        });
-      }
+      keys.forEach(key => {
+        const value = data[key];
+        window.sessionStorage.setItem(key, value !== undefined ? value : "");
+      });
     },
     remove(key: string) {
-      if (key) {
-        window.sessionStorage.remove(key);
-      }
+      if (!isClient || !key) return;
+      window.sessionStorage.removeItem(key);
     },
     clear() {
+      if (!isClient) return;
       window.sessionStorage.clear();
     }
   },
 
   local: {
     get(key: string) {
+      if (!isClient) return null;
       return key ? window.localStorage.getItem(key) : window.localStorage;
     },
     set(data: Record<string, any> = {}) {
+      if (!isClient) return;
       const keys = Object.keys(data);
-      if (keys.length > 0) {
-        keys.forEach(key => {
-          const value = data[key];
-          window.localStorage.setItem(key, value !== undefined ? value : "");
-        });
-      }
+      keys.forEach(key => {
+        const value = data[key];
+        window.localStorage.setItem(key, value !== undefined ? value : "");
+      });
     },
     remove(key: string) {
-      if (key) {
-        window.localStorage.remove(key);
-      }
+      if (!isClient || !key) return;
+      window.localStorage.removeItem(key);
     },
     clear() {
+      if (!isClient) return;
       window.localStorage.clear();
     }
   }
